requirements.txt
flask==1.1.2
google-cloud-vision==2.0.0
gunicorn==20.0.4
python-telegram-bot==13.1
pyyaml==5.3.1